+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from marshpillow import *
+from pillowtalk import *
 
 
 @pytest.fixture(scope="module")
