+++ tests/test_basics/test_session_manager.py
@@ -1,5 +1,5 @@
 import pytest
-from marshpillow import *
+from pillowtalk import *
 
 @pytest.fixture
 def API():
