+++ tests/test_relationships/test_many_to_one.py
@@ -1,5 +1,5 @@
-from marshpillow import *
-from marshpillow.schemas import add_schema
+from pillowtalk import *
+from pillowtalk.schemas import add_schema
 
 
 def test_basic(mybase):
