+++ tests/test_relationships/test_multiple_relationships.py
@@ -1,5 +1,5 @@
 import pytest
-from marshpillow import *
+from pillowtalk import *
