+++ tests/test_with_larger_examples/test_sequence_example/test_examples.py
@@ -1,5 +1,5 @@
 import pytest
-from marshpillow import *
+from pillowtalk import *
 
 # TODO: test when where returns [] or None or [None]
 # TODO: test update() (or autoupdate?)
